新增查询当前登录用户接口

+++ src/main/java/com/doudou/freechat/mapper/UserMapper.java
@@ -11,6 +11,6 @@ public interface UserMapper {
 
     UserDao getUserInfoByName(@Param("userName") String userName);
 
-    long addUser(UserDao record);
+    int addUser(UserDao record);
 
 }
